+++ tests/unit/vcsModule.spec.ts
@@ -1,25 +1,32 @@
-import VcsModule from '../../src/vcsModule.js';
+import { expect } from 'chai';
+import VcsModule, { VcsModuleConfig } from '../../src/vcsModule.js';
 import VcsApp from '../../src/vcsApp.js';
 import VectorLayer from '../../src/layer/vectorLayer.js';
 import Viewpoint from '../../src/util/viewpoint.js';
 import OpenlayersMap from '../../src/map/openlayersMap.js';
 import CesiumMap from '../../src/map/cesiumMap.js';
 
-describe('Module', () => {
-  let app;
-  let module;
-  let startingVp;
+type TestVcsModuleConfig = VcsModuleConfig & {
+  thirdpartyEntry: string;
+};
 
-  before(async () => {
+describe('Module', () => {
+  let app: VcsApp;
+  let module: VcsModule;
+  let startingVp: Viewpoint;
+
+  beforeEach(async () => {
     app = new VcsApp();
     startingVp = new Viewpoint({
       name: 'foo',
       groundPosition: [13, 52],
       distance: 200,
     });
+
     module = new VcsModule({
       name: 'module',
       description: 'description',
+      thirdpartyEntry: 'test',
       layers: [
         new VectorLayer({ name: 'foo' }).toJSON(),
         new VectorLayer({ name: 'bar', activeOnStartup: true }).toJSON(),
@@ -28,14 +35,18 @@ describe('Module', () => {
       maps: [new OpenlayersMap({ name: 'foo' }).toJSON()],
       startingViewpointName: 'foo',
       startingMapName: 'foo',
-    });
+    } as TestVcsModuleConfig);
     await app.addModule(module);
   });
 
-  describe('getting a config', () => {
-    let config;
+  afterEach(() => {
+    app.destroy();
+  });
 
-    before(() => {
+  describe('getting a config', () => {
+    let config: object;
+
+    beforeEach(() => {
       config = module.toJSON();
     });
 
@@ -56,15 +67,24 @@ describe('Module', () => {
   describe('setting a config', () => {
     it('should serialize the current runtime objects for the dynamic module', () => {
       app.setDynamicModule(module);
-      app.layers.getByKey('foo').name = 'fooBar';
+      app.layers.remove(app.layers.getByKey('foo')!);
+      app.layers.add(new VectorLayer({ name: 'fooBar' }));
       app.viewpoints.remove(startingVp);
       app.maps.add(new CesiumMap({ name: 'cesium' }));
       module.setConfigFromApp(app);
       const { config } = module;
       expect(config.layers).to.have.lengthOf(2);
-      expect(!!config.layers.find((l) => l.name === 'fooBar')).to.be.true;
+      expect(!!config.layers?.find((l) => l.name === 'fooBar')).to.be.true;
       expect(config.viewpoints).to.have.lengthOf(1);
       expect(config.maps).to.have.lengthOf(2);
+    });
+
+    it('should not loose keys in the config not defined by the app', () => {
+      module.setConfigFromApp(app);
+      const { config } = module;
+      expect((config as TestVcsModuleConfig).thirdpartyEntry).to.be.equal(
+        'test',
+      );
     });
   });
 });